added skeleton for auto-docstrings
