+++ collie_recs/model/base/__init__.py
@@ -1,0 +1,3 @@
+from collie_recs.model.base.base_pipeline import *
+from collie_recs.model.base.layers import *
+from collie_recs.model.base.trainer import *
