+++ requirements-dev.txt
@@ -18,6 +18,10 @@ alabaster==0.7.12
     # via sphinx
 anyio==3.0.1
     # via jupyter-server
+appnope==0.1.2
+    # via
+    #   ipykernel
+    #   ipython
 argon2-cffi==20.1.0
     # via
     #   jupyter-server
@@ -46,7 +50,7 @@ cachetools==4.2.2
     # via
     #   -c requirements.txt
     #   google-auth
-certifi==2020.12.5
+certifi==2021.5.30
     # via
     #   -c requirements.txt
     #   requests
@@ -67,6 +71,10 @@ decorator==5.0.7
     # via ipython
 defusedxml==0.7.1
     # via nbconvert
+docstring-parser==0.8.1
+    # via
+    #   -c requirements.txt
+    #   collie-recs
 docutils==0.17.1
     # via
     #   m2r2
@@ -85,7 +93,7 @@ flake8==3.9.2
     # via
     #   collie-recs
     #   flake8-docstrings
-fsspec[http]==2021.4.0
+fsspec[http]==2021.5.0
     # via
     #   -c requirements.txt
     #   pytorch-lightning
@@ -97,12 +105,12 @@ google-auth-oauthlib==0.4.4
     # via
     #   -c requirements.txt
     #   tensorboard
-google-auth==1.30.0
+google-auth==1.30.1
     # via
     #   -c requirements.txt
     #   google-auth-oauthlib
     #   tensorboard
-grpcio==1.37.1
+grpcio==1.38.0
     # via
     #   -c requirements.txt
     #   tensorboard
@@ -247,7 +255,7 @@ numpy==1.20.3
     #   tables
     #   tensorboard
     #   torch
-oauthlib==3.1.0
+oauthlib==3.1.1
     # via
     #   -c requirements.txt
     #   requests-oauthlib
@@ -287,7 +295,7 @@ prometheus-client==0.10.1
     #   notebook
 prompt-toolkit==3.0.18
     # via ipython
-protobuf==3.17.0
+protobuf==3.17.2
     # via
     #   -c requirements.txt
     #   tensorboard
@@ -347,7 +355,7 @@ python-dateutil==2.8.1
     #   jupyter-client
     #   matplotlib
     #   pandas
-pytorch-lightning==1.3.1
+pytorch-lightning==1.3.4
     # via
     #   -c requirements.txt
     #   collie-recs
@@ -484,7 +492,7 @@ tornado==6.1
     #   jupyterlab
     #   notebook
     #   terminado
-tqdm==4.60.0
+tqdm==4.61.0
     # via
     #   -c requirements.txt
     #   collie-recs
@@ -507,7 +515,7 @@ typing-extensions==3.10.0.0
     #   -c requirements.txt
     #   aiohttp
     #   torch
-urllib3==1.26.4
+urllib3==1.26.5
     # via
     #   -c requirements.txt
     #   requests
@@ -517,7 +525,7 @@ webencodings==0.5.1
     # via bleach
 websocket-client==0.59.0
     # via jupyter-server
-werkzeug==2.0.0
+werkzeug==2.0.1
     # via
     #   -c requirements.txt
     #   tensorboard
