+++ requirements.txt
@@ -14,25 +14,27 @@ attrs==21.2.0
     # via aiohttp
 cachetools==4.2.2
     # via google-auth
-certifi==2020.12.5
+certifi==2021.5.30
     # via requests
 chardet==4.0.0
     # via
     #   aiohttp
     #   requests
+docstring-parser==0.8.1
+    # via collie-recs (setup.py)
 fire==0.4.0
     # via collie-recs (setup.py)
-fsspec[http]==2021.4.0
+fsspec[http]==2021.5.0
     # via pytorch-lightning
 future==0.18.2
     # via pytorch-lightning
 google-auth-oauthlib==0.4.4
     # via tensorboard
-google-auth==1.30.0
+google-auth==1.30.1
     # via
     #   google-auth-oauthlib
     #   tensorboard
-grpcio==1.37.1
+grpcio==1.38.0
     # via tensorboard
 idna==2.10
     # via
@@ -61,7 +63,7 @@ numpy==1.20.3
     #   tables
     #   tensorboard
     #   torch
-oauthlib==3.1.0
+oauthlib==3.1.1
     # via requests-oauthlib
 packaging==20.9
     # via
@@ -69,7 +71,7 @@ packaging==20.9
     #   torchmetrics
 pandas==1.2.4
     # via collie-recs (setup.py)
-protobuf==3.17.0
+protobuf==3.17.2
     # via tensorboard
 pyasn1-modules==0.2.8
     # via google-auth
@@ -83,7 +85,7 @@ pyparsing==2.4.7
     # via packaging
 python-dateutil==2.8.1
     # via pandas
-pytorch-lightning==1.3.1
+pytorch-lightning==1.3.4
     # via collie-recs (setup.py)
 pytz==2021.1
     # via pandas
@@ -130,7 +132,7 @@ torchmetrics==0.3.2
     # via
     #   collie-recs (setup.py)
     #   pytorch-lightning
-tqdm==4.60.0
+tqdm==4.61.0
     # via
     #   collie-recs (setup.py)
     #   pytorch-lightning
@@ -138,9 +140,9 @@ typing-extensions==3.10.0.0
     # via
     #   aiohttp
     #   torch
-urllib3==1.26.4
+urllib3==1.26.5
     # via requests
-werkzeug==2.0.0
+werkzeug==2.0.1
     # via tensorboard
 wheel==0.36.2
     # via tensorboard
